simpleiot/toolchain/arduino/M5Stack/requirements.txt
colorama==0.4.4
commonmark==0.9.1
importlib-metadata==4.8.2
importlib-resources==5.4.0
packaging==21.2
pkg-about==1.0.3
prompt-toolkit==3.0.22
Pygments==2.10.0
pyinvoke==1.0.4
pyparsing==2.4.7
pyserial==3.5
pyusb==1.2.1
questionary==1.10.0
rich==10.13.0
six==1.16.0
wcwidth==0.2.5
zipp==3.6.0
